DUOS-1068 component refinement, split table out of AdminManageLC component
